# openapi.yaml
...
openapi: 3.0.0
info:
  x-skaffolder-id-project: 5dcba54af1ef4518a5382e85
  title: Wishbox
  description: You can obtain JWT token from login API
  version: 0.1
servers:
  - url: http://localhost:3000/api
    description: Local server for testing
paths:
  /login:
    post:
      x-skaffolder-ignore: true
      summary: Login API
      security: []   # No security
      tags:
        - Security
      requestBody:
        content:
          application/json:
            schema:      # Request body contents
              type: object
              properties:
                username:
                  type: string
                password:
                  type: string
              example:   # Sample object
                username: admin
                password: 62f264d7ad826f02a8af714c0a54b197935b717656b80461686d450f7b3abde4c553541515de2052b9af70f710f0cd8a1a2d3f4d60aa72608d71a63a9a93c0f5
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
  /verifyToken:
    post:
      x-skaffolder-ignore: true
      summary: Verify token API
      security: [] #No security
      tags:
        - Security
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                token:
                  type: string
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
  /changePassword:
    post:
      x-skaffolder-ignore: true
      summary: Change password API
      tags:
        - Security
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                passwordNew:
                  type: string
                passwordOld:
                  type: string
      responses:
        '200':
          description: OK
                

  #-----  Category  -----
...
  /user/:
    post:
      x-skaffolder-id: 5dcba54bf1ef4518a5382e92
      x-skaffolder-name: create
      x-skaffolder-id-resource: 5dcba54af1ef4518a5382e8f
      x-skaffolder-resource: User
      x-skaffolder-crudAction: create
      x-skaffolder-crudType: 
      x-skaffolder-description: CRUD ACTION create
      x-skaffolder-returnDesc: 
      x-skaffolder-returnType: 
      x-skaffolder-url: /
      x-skaffolder-roles: []
      summary: CRUD ACTION create
      tags:
        - User
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                email:
                  type: string
                name:
                  type: string
                password:
                  type: string
                phone:
                  type: string
                roles:
                  type: string
                username:
                  type: string
    get:
      x-skaffolder-id: 5dcba54bf1ef4518a5382e93
      x-skaffolder-name: list
      x-skaffolder-id-resource: 5dcba54af1ef4518a5382e8f
      x-skaffolder-resource: User
      x-skaffolder-crudAction: list
      x-skaffolder-crudType: 
      x-skaffolder-description: CRUD ACTION list
      x-skaffolder-returnDesc: 
      x-skaffolder-returnType: 
      x-skaffolder-url: /
      x-skaffolder-roles: []
      summary: CRUD ACTION list
      tags:
        - User
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/User"
  /user/{id}:
    delete:
      x-skaffolder-id: 5dcba54bf1ef4518a5382e98
      x-skaffolder-name: delete
      x-skaffolder-id-resource: 5dcba54af1ef4518a5382e8f
      x-skaffolder-resource: User
      x-skaffolder-crudAction: delete
      x-skaffolder-crudType: 
      x-skaffolder-description: CRUD ACTION delete
      x-skaffolder-returnDesc: 
      x-skaffolder-returnType: 
      x-skaffolder-url: /{id}
      x-skaffolder-roles: []
      summary: CRUD ACTION delete
      tags:
        - User
      parameters:
        - name: id
          x-skaffolder-type: ObjectId
          in: path
          description: Id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
    get:
      x-skaffolder-id: 5dcba54bf1ef4518a5382e95
      x-skaffolder-name: get
      x-skaffolder-id-resource: 5dcba54af1ef4518a5382e8f
      x-skaffolder-resource: User
      x-skaffolder-crudAction: get
      x-skaffolder-crudType: 
      x-skaffolder-description: CRUD ACTION get
      x-skaffolder-returnDesc: 
      x-skaffolder-returnType: 
      x-skaffolder-url: /{id}
      x-skaffolder-roles: []
      summary: CRUD ACTION get
      tags:
        - User
      parameters:
        - name: id
          x-skaffolder-type: ObjectId
          in: path
          description: Id resource
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
    post:
      x-skaffolder-id: 5dcba54bf1ef4518a5382e97
      x-skaffolder-name: update
      x-skaffolder-id-resource: 5dcba54af1ef4518a5382e8f
      x-skaffolder-resource: User
      x-skaffolder-crudAction: update
      x-skaffolder-crudType: 
      x-skaffolder-description: CRUD ACTION update
      x-skaffolder-returnDesc: 
      x-skaffolder-returnType: 
      x-skaffolder-url: /{id}
      x-skaffolder-roles: []
      summary: CRUD ACTION update
      tags:
        - User
      parameters:
        - name: id
          x-skaffolder-type: ObjectId
          in: path
          description: Id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/User"
    
    
# JWT security
security:
  - bearerAuth: [] 
  
  
# Components
components:
...
  securitySchemes:
    bearerAuth:            # arbitrary name for the security scheme
      type: http
      scheme: bearer
      bearerFormat: JWT    # optional, arbitrary value for documentation purposes

  # Data Model
  schemas:
...
    User:
      x-skaffolder-id: 5dcba54af1ef4518a5382e8f
      x-skaffolder-id-db: 5dcba54af1ef4518a5382e86
      x-skaffolder-id-entity: 5dcba54af1ef4518a5382e87
      x-skaffolder-url: /user
      properties:
        _id:
          type: string
          x-skaffolder-required: true
        email:
          type: string
          x-skaffolder-id-attr: 5dcba54af1ef4518a5382e8a
          x-skaffolder-type: String
        name:
          type: string
          x-skaffolder-id-attr: 5dcba54af1ef4518a5382e8b
          x-skaffolder-type: String
        password:
          type: string
          x-skaffolder-id-attr: 5dcba54af1ef4518a5382e89
          x-skaffolder-type: String
          x-skaffolder-required: true
        phone:
          type: string
          x-skaffolder-id-attr: 5dcbaa9ef1ef4518a5382f19
          x-skaffolder-type: String
          x-skaffolder-unique: 
          x-skaffolder-required: true
        roles:
          type: string
          x-skaffolder-id-attr: 5dcba54af1ef4518a5382e8d
          x-skaffolder-type: String
        username:
          type: string
          x-skaffolder-id-attr: 5dcba54af1ef4518a5382e88
          x-skaffolder-type: String
          x-skaffolder-required: true
      x-skaffolder-relations: 
      required:
        - _id
        - password
        - phone
        - username
      
    
    
  # Databases
...